+++ projects/packages/jetpack-mu-wpcom/src/features/wpcom-site-menu/wpcom-site-menu.php
@@ -435,72 +435,118 @@ function wpcom_maybe_enable_link_manager() {
 add_action( 'init', 'wpcom_maybe_enable_link_manager' );
 
 /**
- * Add the Scheduled Updates menu item to the Plugins menu.
+ * Hides a submenu item.
  *
- * Limited to sites with scheduled updates feature.
+ * Useful in cases where we cannot remove a submenu item because there is external logic
+ * that depends on the route registered by that submenu.
+ *
+ * @param string $menu_slug The slug of the parent menu.
+ * @param string $submenu_slug The slug of the submenu that should be hidden.
  */
-function wpcom_add_scheduled_updates_menu() {
-	// Bail on Simple sites
-	if ( defined( 'IS_WPCOM' ) && IS_WPCOM ) {
+function wpcom_hide_submenu_page( string $menu_slug, string $submenu_slug ) {
+	global $submenu;
+
+	if ( ! isset( $submenu[ $menu_slug ] ) ) {
 		return;
 	}
 
-	/**
-	 * Don't show `Scheduled Updates` to administrators without a WordPress.com account being attached,
-	 * as they don't have access to any of the pages.
-	 */
-	if ( ! current_user_has_wpcom_account() ) {
+	foreach ( $submenu[ $menu_slug ] as $i => $item ) {
+		if ( $submenu_slug !== $item[2] ) {
+			continue;
+		}
+
+		$css_hide_class = 'hide-if-js';
+		$css_classes    = empty( $item[4] ) ? $css_hide_class : $item[4] . ' ' . $css_hide_class;
+
+		// phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
+		$submenu[ $menu_slug ][ $i ][4] = $css_classes;
 		return;
 	}
+}
 
-	// Don't show on staging sites.
-	if ( get_option( 'wpcom_is_staging_site' ) ) {
-		return;
+/**
+ * Handles the Plugins menu for WP.com sites.
+ */
+function wpcom_add_plugins_menu() {
+	global $menu;
+	$is_simple_site          = defined( 'IS_WPCOM' ) && IS_WPCOM;
+	$is_atomic_site          = ! $is_simple_site;
+	$is_nav_redesign_enabled = function_exists( 'wpcom_is_nav_redesign_enabled' ) && wpcom_is_nav_redesign_enabled();
+
+	if ( $is_simple_site ) {
+		$has_plugins_menu = false;
+		foreach ( $menu as &$menu_item ) {
+			if ( 'plugins.php' === $menu_item[2] ) {
+				$has_plugins_menu = true;
+				break;
+			}
+		}
+
+		if ( ! $has_plugins_menu ) {
+			// TODO: Remove `remove_menu_page('plugins.php');` from `/wp-content/admin-plugins/wpcom-misc.php`.
+			add_menu_page(
+				__( 'Plugins', 'jetpack-mu-wpcom' ),
+				__( 'Plugins', 'jetpack-mu-wpcom' ),
+				'manage_options', // Roughly means "is a site admin"
+				'plugins.php',
+				null,
+				'dashicons-admin-plugins',
+				65
+			);
+		}
+
+		if ( function_exists( 'wpcom_plugins_display_marketplace' ) ) {
+			add_submenu_page(
+				'plugins.php',
+				__( 'Add New Plugin', 'jetpack-mu-wpcom' ),
+				__( 'Add New Plugin', 'jetpack-mu-wpcom' ),
+				'manage_options', // Roughly means "is a site admin"
+				'wpcom-install-plugin',
+				'wpcom_plugins_display_marketplace'
+			);
+
+			if ( ! $is_nav_redesign_enabled ) {
+				wpcom_hide_submenu_page( 'plugins.php', 'wpcom-install-plugin' );
+			}
+		}
 	}
 
-	if ( ! function_exists( 'wpcom_site_has_feature' ) ) {
-		return;
-	}
-
-	if ( ! wpcom_site_has_feature( \WPCOM_Features::SCHEDULED_UPDATES ) ) {
+	if ( ! $is_nav_redesign_enabled ) {
 		return;
 	}
 
 	$domain = wp_parse_url( home_url(), PHP_URL_HOST );
-
 	add_submenu_page(
 		'plugins.php',
-		esc_attr__( 'Scheduled Updates', 'jetpack-mu-wpcom' ),
-		__( 'Scheduled Updates', 'jetpack-mu-wpcom' ),
-		'update_plugins',
-		esc_url( "https://wordpress.com/plugins/scheduled-updates/$domain" ),
+		/* translators: Name of the Plugins submenu that links to the Plugins Marketplace */
+		__( 'Marketplace', 'jetpack-mu-wpcom' ),
+		/* translators: Name of the Plugins submenu that links to the Plugins Marketplace */
+		__( 'Marketplace', 'jetpack-mu-wpcom' ),
+		'manage_options', // Roughly means "is a site admin"
+		'https://wordpress.com/plugins/' . $domain,
 		null
 	);
-}
-add_action( 'admin_menu', 'wpcom_add_scheduled_updates_menu' );
 
-/**
- * Add the Plugins menu item to the admin menu on simple sites.
- */
-function wpcom_add_plugins_menu() {
-
-	if ( ! defined( 'IS_WPCOM' ) || ! IS_WPCOM ) {
-		return;
-	}
-
-	if ( function_exists( 'wpcom_is_nav_redesign_enabled' ) && wpcom_is_nav_redesign_enabled() ) {
-		$domain              = wp_parse_url( home_url(), PHP_URL_HOST );
-		$can_install_plugins = function_exists( 'wpcom_site_has_feature' ) && wpcom_site_has_feature( WPCOM_Features::INSTALL_PLUGINS );
-
-		add_menu_page(
-			__( 'Plugins', 'jetpack-mu-wpcom' ),
-			__( 'Plugins', 'jetpack-mu-wpcom' ),
-			'manage_options', // Roughly means "is a site admin"
-			$can_install_plugins ? 'https://wordpress.com/plugins/' . $domain : 'plugins.php',
-			null,
-			'dashicons-admin-plugins',
-			65
-		);
+	if ( $is_atomic_site ) {
+		if (
+			/**
+			 * Don't show `Scheduled Updates` to administrators without a WordPress.com account being attached,
+			 * as they don't have access to any of the pages.
+			 */
+			current_user_has_wpcom_account() &&
+			! get_option( 'wpcom_is_staging_site' ) &&
+			function_exists( 'wpcom_site_has_feature' ) &&
+			wpcom_site_has_feature( \WPCOM_Features::SCHEDULED_UPDATES )
+		) {
+			add_submenu_page(
+				'plugins.php',
+				esc_attr__( 'Scheduled Updates', 'jetpack-mu-wpcom' ),
+				__( 'Scheduled Updates', 'jetpack-mu-wpcom' ),
+				'update_plugins',
+				esc_url( "https://wordpress.com/plugins/scheduled-updates/$domain" ),
+				null
+			);
+		}
 	}
 }
 add_action( 'admin_menu', 'wpcom_add_plugins_menu' );