+++ projects/plugins/social/src/js/components/header/index.js
@@ -33,11 +33,13 @@ const Header = () => {
 		siteSuffix,
 		blogID,
 		showShareLimits,
+		hasPaidFeatures,
+		useAdminUiV1,
 	} = useSelect( select => {
 		const store = select( socialStore );
 		return {
 			connectionsAdminUrl: connectionData.adminUrl,
-			hasConnections: Object.keys( connectionData.connections || {} ).length > 0,
+			hasConnections: store.getConnections().length > 0,
 			isModuleEnabled: store.isModuleEnabled(),
 			newPostUrl: `${ store.getAdminUrl() }post-new.php`,
 			postsCount: store.getSharedPostsCount(),
@@ -45,6 +47,8 @@ const Header = () => {
 			siteSuffix: store.getSiteSuffix(),
 			blogID: store.getBlogID(),
 			showShareLimits: store.showShareLimits(),
+			hasPaidFeatures: store.hasPaidFeatures() || store.hasPaidPlan(),
+			useAdminUiV1: store.useAdminUiV1(),
 		};
 	} );
 	const { hasConnectionError } = useConnectionErrorNotice();
@@ -72,7 +76,7 @@ const Header = () => {
 				<Col sm={ 4 } md={ 4 } lg={ 5 }>
 					<H3 mt={ 2 }>{ __( 'Write once, post everywhere', 'jetpack-social' ) }</H3>
 					<div className={ styles.actions }>
-						{ isModuleEnabled && ! hasConnections && (
+						{ ! useAdminUiV1 && isModuleEnabled && ! hasConnections && (
 							<Button href={ connectionsAdminUrl } isExternalLink={ true }>
 								{ __( 'Connect accounts', 'jetpack-social' ) }
 							</Button>
@@ -102,19 +106,6 @@ const Header = () => {
 									) }
 								</i>
 							</Text>
-							<ContextualUpgradeTrigger
-								className={ styles.cut }
-								description={ __(
-									'Unlock unlimited shares and advanced posting options',
-									'jetpack-social'
-								) }
-								cta={ __( 'Get a Jetpack Social Plan', 'jetpack-social' ) }
-								href={ getRedirectUrl( 'jetpack-social-admin-page-upsell', {
-									site: blogID ?? siteSuffix,
-									query: 'redirect_to=admin.php?page=jetpack-social',
-								} ) }
-								tooltipText={ __( 'Share as a post for more engagement', 'jetpack-social' ) }
-							/>
 						</>
 					) : (
 						<StatCards
@@ -134,6 +125,18 @@ const Header = () => {
 							] }
 						/>
 					) }
+					{ ! hasPaidFeatures ? (
+						<ContextualUpgradeTrigger
+							className={ styles.cut }
+							description={ __( 'Unlock advanced posting options', 'jetpack-social' ) }
+							cta={ __( 'Get a Jetpack Social Plan', 'jetpack-social' ) }
+							href={ getRedirectUrl( 'jetpack-social-admin-page-upsell', {
+								site: blogID ?? siteSuffix,
+								query: 'redirect_to=admin.php?page=jetpack-social',
+							} ) }
+							tooltipText={ __( 'Share as a post for more engagement', 'jetpack-social' ) }
+						/>
+					) : null }
 				</Col>
 			</Container>
 		</>